Testing support for windows prelude load
